TST: test_accumulate
