+++ shed/tests/test_streams.py
@@ -17,12 +17,12 @@
 from itertools import zip_longest
 
 import pytest
-from numpy.testing import assert_allclose, assert_equal
+from numpy.testing import assert_equal
 from streamz.core import Stream
 
 import shed.event_streams as es
 from shed.tests.utils import SinkAssertion
-from ..event_streams import dstar, star
+from ..event_streams import star
 from ..utils import to_event_model
 
 
@@ -34,265 +34,6 @@ def test_clear():
     assert s.md == {'hello': 'globe'}
     s._clear()
     assert s.md == {'hello': 'world'}
-
-
-def test_scan(exp_db, start_uid1):
-    source = Stream()
-
-    def add(img1, img2):
-        return img1 + img2
-
-    dp = es.accumulate(dstar(add), source,
-                       state_key='img1',
-                       input_info={'img2': 'pe1_image'},
-                       output_info=[('img', {
-                           'dtype': 'array',
-                           'source': 'testing'})])
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    state = None
-    for o, l in zip(exp_db.restream(ih1, fill=True), L):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            if state is None:
-                state = o[1]['data']['pe1_image']
-            else:
-                state += o[1]['data']['pe1_image']
-            assert_allclose(state, l[1]['data']['img'])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-@pytest.mark.xfail(raises=TypeError)
-def test_scan_fail(exp_db, start_uid1):
-    source = Stream()
-
-    def add(img1, img2):
-        return img1 + img2
-
-    dp = es.accumulate(dstar(add), source,
-                       state_key='i',
-                       input_info={'i': 'pe1_image'},
-                       output_info=[('img', {
-                           'dtype': 'array',
-                           'source': 'testing'})])
-    sa = SinkAssertion()
-    sa.expected_docs = {'start', 'descriptor', 'event', 'stop'}
-    dp.sink(star(sa))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-
-@pytest.mark.xfail(raises=ValueError)
-def test_scan_fail_input_info(exp_db, start_uid1):
-    ''' should fail on bad input info (more than one input).'''
-    source = Stream()
-
-    # just test this raises ValueError
-    es.accumulate(dstar(lambda x: x), source,
-                  state_key='i',
-                  input_info={'i': 'pe1_image', 'i2': 'pe2_image'},
-                  output_info=[('img', {
-                               'dtype': 'array',
-                               'source': 'testing'})])
-
-
-def test_scan_start_func(exp_db, start_uid1):
-    source = Stream()
-
-    def add(img1, img2):
-        return img1 + img2
-
-    def get_array(img2):
-        return img2
-
-    dp = es.accumulate(dstar(add), source,
-                       start=dstar(get_array),
-                       state_key='img1',
-                       input_info={'img2': 'pe1_image'},
-                       output_info=[('img', {
-                           'dtype': 'array',
-                           'source': 'testing'})])
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    state = None
-    for o, l in zip(exp_db.restream(ih1, fill=True), L):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            if state is None:
-                state = o[1]['data']['pe1_image']
-            else:
-                state += o[1]['data']['pe1_image']
-            assert_allclose(state, l[1]['data']['img'])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_scan_full_event(exp_db, start_uid1):
-    source = Stream()
-
-    def add(i, j):
-        return i + j
-
-    dp = es.accumulate(dstar(add), source,
-                       state_key='i',
-                       input_info={'j': 'seq_num'},
-                       output_info=[('total', {
-                           'dtype': 'int',
-                           'source': 'testing'})],
-                       full_event=True)
-
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    state = None
-    for o, l in zip(exp_db.restream(ih1, fill=True), L):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            if state is None:
-                state = o[1]['seq_num']
-            else:
-                state += o[1]['seq_num']
-            assert_allclose(state, l[1]['data']['total'])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_scan_multi_header_False(exp_db, start_uid1):
-    source = Stream()
-
-    def add(img1, img2):
-        return img1 + img2
-
-    dp = es.accumulate(dstar(add), source,
-                       state_key='img1',
-                       input_info={'img2': 'pe1_image'},
-                       output_info=[('img', {
-                           'dtype': 'array',
-                           'source': 'testing'})])
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    state = None
-    for o, l in zip(exp_db.restream(ih1, fill=True), L):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            if state is None:
-                state = o[1]['data']['pe1_image']
-            else:
-                state += o[1]['data']['pe1_image']
-            assert_allclose(state, l[1]['data']['img'])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    state = None
-    for o, l in zip(exp_db.restream(ih1, fill=True), L):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            if state is None:
-                state = o[1]['data']['pe1_image']
-            else:
-                state += o[1]['data']['pe1_image']
-            assert_allclose(state, l[1]['data']['img'])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_scan_multi_header_True(exp_db, start_uid1):
-    source = Stream()
-
-    def add(img1, img2):
-        return img1 + img2
-
-    dp = es.accumulate(dstar(add), source,
-                       state_key='img1',
-                       input_info={'img2': 'pe1_image'},
-                       output_info=[('img', {
-                           'dtype': 'array',
-                           'source': 'testing'})],
-                       across_start=True)
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    state = None
-    for o, l in zip(exp_db.restream(ih1, fill=True), L):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            if state is None:
-                state = o[1]['data']['pe1_image']
-            else:
-                state += o[1]['data']['pe1_image']
-            assert_allclose(state, l[1]['data']['img'])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-    L.clear()
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for o, l in zip(exp_db.restream(ih1, fill=True), L):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            state += o[1]['data']['pe1_image']
-            assert_allclose(state, l[1]['data']['img'])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
 
 
 def test_zip(exp_db, start_uid1, start_uid3):
